+++ netbox_more_metrics/migrations/0002_metric_metric_value.py
@@ -1,0 +1,15 @@
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+    dependencies = [
+        ("netbox_more_metrics", "0001_initial"),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name="metric",
+            name="metric_value",
+            field=models.CharField(default="count", max_length=50),
+        ),
+    ]