+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0 (2023-03-09)
+
+### Features
+* [#3](https://github.com/TheDJVG/netbox-more-metrics/issues/5) - Metric value based on NetBox objects.
+  * For example get the percentage of IPs in a Prefix/Aggregate occupied, or how full a rack is.
+
 ## 0.1.1 (2023-02-27)
 
 ### Bugfixes